+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # 3.1.0
 
+* Added config provider.
+* Added cacheEnabled option to config provider.
 * `drf.loadItem` and `drf.loadList` now reject the deferred object on errors.
 
 # 3.0.1